requirements.txt
Django==2.2.1
djangorestframework==3.9.4
# django-smart-selects==1.5.4
# git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
git+https://github.com/kajogo777/django-smart-selects.git@master
django-jet==1.0.8
django-filter==2.1.0
reportlab==3.5.23